+++ lib/src/widget/card_swiper.dart
@@ -141,6 +141,20 @@ class CardSwiper extends StatefulWidget {
   /// The widget is automatically faded with `progress`.
   final CardSwiperOverlayBuilder? overlayBuilder;
 
+  /// The direction of the swipe to trigger an undo.
+  /// Defaults to UndoDirection.left.
+  final UndoDirection undoDirection;
+
+  /// If true, the previous card will be shown in the background when swiping in the undo direction,
+  /// and the swipe action in the direction specified by [undoDirection] will trigger an undo.
+  /// Defaults to false.
+  final bool showBackCardOnUndo;
+
+  /// The horizontal swipe distance threshold, in pixels, after which the back card
+  /// switches to the previous card. Only active when [showBackCardOnUndo] is true.
+  /// Defaults to 50.0.
+  final double undoSwipeThreshold;
+
   const CardSwiper({
     required this.cardBuilder,
     required this.cardsCount,
@@ -164,7 +178,9 @@ class CardSwiper extends StatefulWidget {
     this.numberOfCardsDisplayed = 2,
     this.onUndo,
     this.backCardOffset = const Offset(0, 40),
-    this.overlayBuilder,
+    this.undoDirection = UndoDirection.left,
+    this.showBackCardOnUndo = false,
+    this.undoSwipeThreshold = 50.0,
     super.key,
   })  : assert(
           maxAngle >= 0 && maxAngle <= 360,
@@ -185,6 +201,10 @@ class CardSwiper extends StatefulWidget {
         assert(
           initialIndex >= 0 && initialIndex < cardsCount,
           'initialIndex must be between 0 and [cardsCount]',
+        ),
+        assert(
+          undoSwipeThreshold >= 0,
+          'undoSwipeThreshold must be a positive value',
         );
 
   @override