+++ lib/src/widget/card_swiper_state.dart
@@ -12,7 +12,10 @@ class _CardSwiperState<T extends Widget> extends State<CardSwiper>
   bool _tappedOnTop = false;
 
   final _undoableIndex = Undoable<int?>(null);
+
   final Queue<CardSwiperDirection> _directionHistory = Queue();
+
+  int? _backgroundCardIndex;
 
   int? get _currentIndex => _undoableIndex.state;
 
@@ -43,6 +46,14 @@ class _CardSwiperState<T extends Widget> extends State<CardSwiper>
       initialOffset: widget.backCardOffset,
       onSwipeDirectionChanged: onSwipeDirectionChanged,
     );
+  }
+
+  @override
+  void didUpdateWidget(CardSwiper oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    controllerSubscription?.cancel();
+    controllerSubscription =
+        widget.controller?.events.listen(_controllerListener);
   }
 
   void onSwipeDirectionChanged(CardSwiperDirection direction) {
@@ -173,13 +184,29 @@ class _CardSwiperState<T extends Widget> extends State<CardSwiper>
         },
         onPanUpdate: (tapInfo) {
           if (!widget.isDisabled) {
-            setState(
-              () => _cardAnimation.update(
+            setState(() {
+              _cardAnimation.update(
                 tapInfo.delta.dx,
                 tapInfo.delta.dy,
                 _tappedOnTop,
-              ),
-            );
+              );
+
+              if (widget.showBackCardOnUndo) {
+                final shouldShowPreviousCard = switch (widget.undoDirection) {
+                  UndoDirection.left =>
+                    _cardAnimation.left < -widget.undoSwipeThreshold,
+                  UndoDirection.right =>
+                    _cardAnimation.left > widget.undoSwipeThreshold,
+                };
+
+                if (shouldShowPreviousCard) {
+                  _backgroundCardIndex =
+                      _undoableIndex.previousState ?? getValidIndexOffset(-1);
+                } else {
+                  _backgroundCardIndex = _nextIndex;
+                }
+              }
+            });
           }
         },
         onPanEnd: (tapInfo) {
@@ -211,6 +238,18 @@ class _CardSwiperState<T extends Widget> extends State<CardSwiper>
   }
 
   Widget _backItem(BoxConstraints constraints, int index) {
+    int? cardIndex;
+
+    if (widget.showBackCardOnUndo && index == 1) {
+      cardIndex = _backgroundCardIndex ?? getValidIndexOffset(index);
+    } else {
+      cardIndex = getValidIndexOffset(index);
+    }
+
+    if (cardIndex == null) {
+      return const SizedBox.shrink();
+    }
+
     return Positioned(
       top: (widget.backCardOffset.dy * index) - _cardAnimation.difference.dy,
       left: (widget.backCardOffset.dx * index) - _cardAnimation.difference.dx,
@@ -218,7 +257,7 @@ class _CardSwiperState<T extends Widget> extends State<CardSwiper>
         scale: _cardAnimation.scale - ((1 - widget.scale) * (index - 1)),
         child: ConstrainedBox(
           constraints: constraints,
-          child: widget.cardBuilder(context, getValidIndexOffset(index)!, 0, 0),
+          child: widget.cardBuilder(context, cardIndex, 0, 0),
         ),
       ),
     );
@@ -267,6 +306,9 @@ class _CardSwiperState<T extends Widget> extends State<CardSwiper>
   void _reset() {
     onSwipeDirectionChanged(CardSwiperDirection.none);
     _detectedDirection = CardSwiperDirection.none;
+
+    _backgroundCardIndex = null;
+
     setState(() {
       _animationController.reset();
       _cardAnimation.reset();
@@ -358,10 +400,13 @@ class _CardSwiperState<T extends Widget> extends State<CardSwiper>
     if (_currentIndex == null) return null;
 
     final index = _currentIndex! + offset;
+
     if (!widget.isLoop && !index.isBetween(0, widget.cardsCount - 1)) {
       return null;
     }
-    return index % widget.cardsCount;
+
+    final result = index % widget.cardsCount;
+    return result < 0 ? result + widget.cardsCount : result;
   }
 
   Future<void> showSwipeOptionsDialog(